services/api_service/pyproject.toml
[tool.poetry]
authors = ["Sylvain Lesage <sylvain.lesage@huggingface.co>"]
description = "REST API app"
name = "api_service"
version = "0.1.0"

[tool.poetry.dependencies]
datasets_preview_backend = { path = "../datasets_preview_backend", develop = false }
python = "3.9.6"
python-dotenv = "^0.19.1"
starlette = "^0.16.0"
uvicorn = "^0.14.0"
watchdog = { extras = ["watchmedo"], version = "^2.1.3" }

[tool.poetry.dev-dependencies]
bandit = "^1.7.0"
black = "^22.1.0"
flake8 = "^3.9.2"
isort = "^5.9.3"
mypy = "0.812"
poetryup = "^0.3.8"
pytest = "^6.2.5"
pytest-cov = "^2.12.1"
safety = "^1.10.3"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["api_service"]

[tool.isort]
profile = "black"

[tool.black]
line-length = 119
preview = true

[tool.mypy]
strict = true
